+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 3.4.0
+
+### Minor Changes
+
+-   Implemented new Meteogram Widget
+-   Created new Forecast page
+-   Refactoring WeatherIcon, added getWeatherIconUrl function
+-   Refactoring WidgetChart and WidgetForecastTable
+-   Columns for WidgetForecastTable are moved to the component, a set of columns is created
+-   Created getSampledData UI function
+-   Updated README, CHANGELOG and fixed styles
+-   Updated UI libraries
+
 ## 3.3.4
 
 ### Path Changes